+++ lib/src/data/data_barrel.dart
@@ -1,0 +1,1 @@
+export 'google_places_api.dart';
