+++ lib/src/model/model_barrel.dart
@@ -1,0 +1,2 @@
+export 'place_details.dart';
+export 'prediction.dart';
