+++ lib/src/model/place_details.dart
