+++ lib/src/model/prediction.dart
