+++ lib/src/src_barrel.dart
@@ -1,0 +1,3 @@
+export 'data/data_barrel.dart';
+export 'model/model_barrel.dart';
+export 'widgets/widgets_barrel.dart';
