+++ lib/src/widgets/widgets_barrel.dart
@@ -1,0 +1,1 @@
+export 'google_places_autocomplete_text_field.dart';
